fix(chat): Remove username parameter from getChatStream (#161)
Removes the `username` parameter from the `getChatStream` method in the
`AiChatService` class. The `username` parameter was previously used to set
the `CHAT_MEMORY_CONVERSATION_ID_KEY` parameter in the chat client, but
this has been changed to use the `exerciseSlug` instead. This change
simplifies the method signature and removes the need to pass the
`Authentication` object from the controller to the service.