feat(roadmaps): enroll funcionality (#165)
* feat(education): Add display order to exercise info

Adds a new `displayOrder` field to the `ExerciseInfo` DTO to represent the order in which the exercises should be displayed. This change is necessary to ensure the exercises are presented to the user in the correct order.

* feat(education): add chrono dependency

The changes in this commit add the `chrono` dependency to the `education` crate in the Tauri desktop application. This dependency is required to handle date and time-related functionality within the application.

* feat(roadmaps): add enroll_in_roadmap command

This commit adds a new Tauri command `enroll_in_roadmap` to the education module. This command allows users to enroll in a specific roadmap by providing the roadmap ID.

The changes include:

- Adding a new `RoadmapEnrollmentResponse` struct to the `education` module to represent the response from the enrollment API.
- Implementing the `enroll_in_roadmap` function in the `roadmaps` module to handle the enrollment process.
- Registering the new `enroll_in_roadmap` command in the `run` function of the main `lib.rs` file.

These changes enable the desktop application to provide users with the ability to enroll in roadmaps, which is an important feature for the education platform.