feat: bump desktop version (#172)
* feat(release): Bump version to 0.1.6 and update license

The changes in this commit include:

1. Bump the version number from 0.1.5 to 0.1.6 in the Cargo.toml and tauri.conf.json files.
2. Update the license in the Cargo.toml file from an empty string to "AGPL-3".
3. Add the "license" and "licenseFile" fields to the "bundle" section in the tauri.conf.json file, setting the license to "AGPL-3.0" and the license file to "../../../LICENSE".

These changes are necessary to reflect the updated version and license information for the Cortex Desktop Application.

* feat(ai_chat): Add safety rating and usage metadata structs

The changes in this commit add two new structs, `SafetyRating` and `UsageMetadata`, to the `gemini.rs` file. These structs are used to deserialize the corresponding data from the Gemini API response. The `#[allow(dead_code)]` attribute is added to these structs to suppress the compiler warning about unused fields.