feat(api): add forgot and reset password routes
The changes in this commit add two new API routes to the `API_ROUTES` object:

1. `FORGOT_PASSWORD`: `${API_BASE_URL}/auth/forgot-password`
2. `RESET_PASSWORD`: `${API_BASE_URL}/auth/reset-password`

These routes are added to support the functionality of password reset and password recovery for the application.

feat(roadmap): add display order to roadmap course

This commit adds a new property `displayOrder` to the `RoadmapCourseDTO` class. This property will be used to determine the order in which the courses are displayed in the roadmap.

The `RoadmapMapper` interface is also updated to include the `displayOrder` property when mapping the `RoadmapCourseDTO` objects.