chore: add terms and conditions (#186)
* feat: add footer content and structure

This commit adds new content and structure to the footer component in the frontend application.

The main changes are:

- Adds three new data objects (`companyItems`, `productItems`, `legalItems`) to store the footer links and their corresponding text and URLs.
- Updates the `FooterSection` component to accept an array of objects with `text` and `link` properties, instead of just an array of strings.
- Replaces the hardcoded footer links with the new data objects in the `AppFooter` component.
- Updates the `CompanyInfo` component to display the correct company address.

These changes improve the maintainability and flexibility of the footer by separating the content from the structure, making it easier to update the footer links in the future.

* feat(terms-and-conditions): Add new terms and conditions page

The new terms and conditions page provides a comprehensive overview of the Cortex platform's general information, privacy and data policies, platform usage, and legal aspects. This page aims to ensure users are well-informed about the platform's terms and conditions before using the service.