add-mentor-details-to-roadmap-overview (#196)
* 🚀 feat(useRoadmaps): add enrollments to exported functions

Adds the `enrollments` function to the exported functions from the
`useRoadmaps` composable. This allows the `enrollments` function to be
used in other parts of the application when working with roadmaps.

apps/desktop/src/composables/useRoadmaps.ts

🎨 feat(explore-page): display mentor information on roadmap card

Adds the `mentor` property to the `RoadmapCard` component to display the
mentor information for the roadmap. This provides users with more
detailed information about the roadmap they are viewing.

apps/desktop/src/pages/explore/[slug]/index.vue

* ✨ feat(roadmap): add mentor details to roadmap overview

Adds the mentor details to the roadmap overview component. This includes
the mentor's full name, username, and avatar image. This information is
now displayed in the roadmap overview to provide more context about the
roadmap's mentor.