update-cover-image-and-remove-explore-exercises-link (#198)
* ✨ feat(index): update cover image and remove explore exercises

Updates the cover image on the index page to a local asset and removes the "Explorar Ejercicios" action from the common actions list.

The cover image is now loaded from a local asset, which improves the performance and reliability of the application. The "Explorar Ejercicios" action was removed as it was not a critical feature and helped to simplify the user interface.

* ✨ feat(roadmap): Optimize roadmap progress and caching

Improve the performance and caching of roadmap progress and roadmap
data. Remove unnecessary caching annotations and optimize the
`calculateRoadmapProgress` method to avoid redundant database queries.
Also, add a new Redis data source configuration to the project.